Get WPP data on an indicator from a specific location

Get data for a specific indicator, for a specific location from the WPP API You will need to get the ID of indicators and locations from get_id to pass into this.

Usage

get_indicator_data(
  indicator_id,
  location_id,
  start_year = 1950,
  end_year = 2100,
  .progress = TRUE
)

Arguments

indicator_id

Numeric ID for the indicator. Common choices include:

46: Age by sex in 5 year age groups

47: Age by sex in 1 year age groups

location_id

Numeric ID for the location (country)

start_year

Numeric. Year to start estimates. Minimum 1950, maximum 2100. Default 1950.

end_year

Numeric. Year to end estimates. Minimum 1951, maximum 2100. Default 2100.

.progress

Boolean. Whether to show progress bars for query. Default TRUE.

Value

tibble with the following columns:

locationId: Numeric ID of location

location: Description of location

iso3: 3-character country code of location

iso2: 2-character country code of location

locationTypeId:

indicatorId: Numeric ID of indicator

indicator: Description of indicator

indicatorDisplayName: Publishable display name of indicator

Details

Note: This function looks for your WPP API key in the environment variable WPP_API_KEY. You can set this using set_wpp_api_key or in your .Renviron file.

Examples

aus_id <- get_id("Australia", type = "locations", search = FALSE, .progress = FALSE)
age_id <- get_id("Population by 5-year age groups and sex", type = "Indicators", search = FALSE, .progress = FALSE)
get_indicator_data(indicator_id = age_id$id, location_id = aus_id$id, start_year = 2020, end_year = 2024)
#> Error in check_wpp_api_key(): No API key detected. Set using `set_wpp_key_key()` or inside `.Renviron`